refactor: 统一 formatDuration 函数到 utils/datetime.ts
- 将 TodoDetail.tsx 和 ExecutionPanel.tsx 中的 formatDuration 合并到 utils/datetime.ts
- 各组件统一从 datetime.ts 导入，不再重复定义
- 删除冗余代码约 30 行
