+++ frontend/src/components/TodoDetail.tsx
@@ -11,22 +11,13 @@ import { parseLogsToMessages } from './ChatView';
 import { TimelineFlow } from './TimelineFlow';
 import type { TimelineRecord } from './TimelineFlow';
 import * as db from '../utils/database';
-import { formatLocalDateTime } from '../utils/datetime';
+import { formatLocalDateTime, formatDuration } from '../utils/datetime';
 import { conversationToYaml } from '../utils/markdown';
 import { AnimatedNumber } from './AnimatedNumber';
 import { getExecutorOption, supportsResume } from '../types';
 import { ExecutorBadge } from './ExecutorBadge';
 import XMarkdown from '@ant-design/x-markdown';
 import type { ExecutionSummary, Todo, TodoItem, ExecutionRecord, ExecutionStats, LogEntry } from '../types';
-
-/** 格式化耗时为人类可读格式，最多4位: 1h2m */
-function formatDuration(seconds: number): string {
-  const h = Math.floor(seconds / 3600);
-  const m = Math.floor((seconds % 3600) / 60);
-  if (h > 0) return `${h}h${m}m`;
-  if (m > 0) return `${m}m`;
-  return `${seconds}s`;
-}
 
 /** 计算从 started_at 到现在的 elapsed time (秒) */
 function getElapsedSeconds(startedAt: string): number {
